Responsive 404 added
